Add files via upload
Adding and deleting tasks from a to-do list is a common and essential feature in many productivity apps and tools. Here's a brief description of how these actions work:

Adding a task: To add a new task to your to-do list, you typically need to click on a button or icon that says "Add task" or something similar. Then, you can type in the name of the task and optionally set a due date, assign it to a project or category, and add any notes or details that you want to remember about it. Once you've entered all the information, you can save the task to your list.

Deleting a task: To delete a task from your to-do list, you usually need to select the task by clicking on it or checking a box next to it. Then, you can click on a "Delete" or "Remove" button to remove the task from your list. Some apps may also allow you to delete a task by swiping it to the side or dragging it to a trash can icon.

It's worth noting that some to-do list apps also allow you to edit or modify existing tasks, such as changing the due date or adding additional notes. These features can help you stay organized and on top of your tasks and responsibilities.